Refactored Handlers - Moved handlers to separate package
